VMS-12526: Enable analytics object detection on cameras by default
VMS-12865: Link to Server API returned to Generic Event settings widget
VMS-9844: Use more explicit option when installing a package with dpkg

<==== Mercurial data ====>
Branch: vms_4.0

+++ libs/nx_utils/src/nx/utils/test_support/test_with_temporary_directory.cpp
@@ -42,8 +42,11 @@ TestWithTemporaryDirectory::TestWithTemporaryDirectory(
 
 TestWithTemporaryDirectory::~TestWithTemporaryDirectory()
 {
-    const bool removed = m_tmpDir.removeRecursively();
-    NX_ASSERT(removed);
+    if (!TestOptions::keepTemporaryDirectory())
+    {
+        const bool removed = m_tmpDir.removeRecursively();
+        NX_ASSERT(removed);
+    }
 }
 
 QString TestWithTemporaryDirectory::testDataDir() const
